+++ lib/shipcloud/operations/all.rb
@@ -3,9 +3,12 @@ module Shipcloud
     module All
       module ClassMethods
         # Loads all Objects of the resource
-        def all
-          response = Shipcloud.request(:get, base_url, {})
-          response.map {|hash| self.new(hash) }
+        def all(filter = {})
+          response = Shipcloud.request(:get, base_url, filter)
+          if index_response_root
+            response = response.fetch(index_response_root, [])
+          end
+          response.map { |hash| new(hash) }
         end
       end