+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - This CHANGELOG file (Following "[Keep a CHANGELOG](http://keepachangelog.com/)")
 - Create, find, update and index operations for address resource. (#4)
 - Services attribute to carriers call (#5)
+- Index operation for shipment resources with optional filter parameters. (#6)
 - Create, find and index operations for webhook resource. (#7)
 - Added the following ruby versions to travis-ci test runs:
   - 2.1.7